// src/utils/types.ts
export enum CLIMATES {
    Thunderstorm = 'Thunderstorm',
    Drizzle = 'Drizzle',
    Rain = 'Rain',
    Snow = 'Snow',
    Mist = 'Mist',
    Smoke = 'Smoke',
    Haze = 'Haze',
    Dust = 'Dust',
    Fog = 'Fog',
    Sand = 'Sand',
    Ash = 'Ash',
    Squall = 'Squall',
    Tornado	= 'Tornado',
    Clear = 'Clear',
    Clouds = 'Clouds'
};